+++ src/main/java/net/enzo/plus/common/item/tools/ItemInfinityAxe.java
@@ -10,10 +10,7 @@ import net.minecraft.entity.Entity;
 import net.minecraft.entity.EntityLivingBase;
 import net.minecraft.entity.effect.EntityLightningBolt;
 import net.minecraft.entity.player.EntityPlayer;
-import net.minecraft.item.EnumRarity;
-import net.minecraft.item.Item;
-import net.minecraft.item.ItemAxe;
-import net.minecraft.item.ItemStack;
+import net.minecraft.item.*;
 import net.minecraft.util.EntityDamageSource;
 import net.minecraft.world.World;
 import net.minecraftforge.common.util.EnumHelper;
@@ -26,6 +23,11 @@ public class ItemInfinityAxe extends ItemAxe {
         setTextureName("plus:infinity_axe");
         setCreativeTab(PlusMod.tab);
         setMaxDamage(0);
+    }
+
+    @Override
+    public EnumAction getItemUseAction(ItemStack p_77661_1_) {
+        return EnumAction.block;
     }
     @Override
     public boolean hitEntity(ItemStack stack, EntityLivingBase victim, EntityLivingBase player) {