+++ src/main/java/net/enzo/plus/common/item/tools/ItemInfinityPickaxe.java
@@ -12,10 +12,7 @@ import net.minecraft.entity.Entity;
 import net.minecraft.entity.EntityLivingBase;
 import net.minecraft.entity.effect.EntityLightningBolt;
 import net.minecraft.entity.player.EntityPlayer;
-import net.minecraft.item.EnumRarity;
-import net.minecraft.item.Item;
-import net.minecraft.item.ItemPickaxe;
-import net.minecraft.item.ItemStack;
+import net.minecraft.item.*;
 import net.minecraft.util.EntityDamageSource;
 import net.minecraft.world.World;
 import net.minecraftforge.common.util.EnumHelper;
@@ -31,6 +28,11 @@ public class ItemInfinityPickaxe extends ItemPickaxe {
         setTextureName("plus:infinity_pickaxe");
         setCreativeTab(PlusMod.tab);
         setMaxDamage(0);
+    }
+
+    @Override
+    public EnumAction getItemUseAction(ItemStack p_77661_1_) {
+        return EnumAction.block;
     }
     @Override
     public boolean hitEntity(ItemStack stack, EntityLivingBase victim, EntityLivingBase player) {